requirements.txt
bcrypt==4.3.0
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
pymongo==4.12.0
Werkzeug==3.1.3
python-dotenv
